Rewrite to split out decoration and service
Splitting out a service as a default means module doesn't have to be
loaded at application bootstrap time in order to run a decorator. Plus
decoration is gross. People who want to decorate still have the old
functionality but by default it's just a simple service w/ a $q dep.

https://github.com/bendrucker/angular-stripe/issues/49#issuecomment-242278905